Consolidate modules into /learn/[id] route, remove projects and skills sub-routes
