+++ TASKS.md
@@ -19,7 +19,7 @@
 
 ## Sprint 3: First teaching module with /learn dashboard
 
-- [ ] add /learn route and dashboard ui rendering first teaching module, #19
+- [x] add /learn route and dashboard ui rendering first teaching module, #19
 - [ ] define module schema and author module content for simple spa app, #20
 - [ ] wire selection to load module files into the editor and file tree, #21
 - [ ] add App Router navigation between editor and dashboard interfaces, #22
